+++ runtime/assembly/modules/balances/index.ts
@@ -1,0 +1,3 @@
+export * from "./account-data";
+export * from "./account-id";
+export * from "./balances-module";
